Pull out Printf and VerificationStatement macros
Update macros to compile in scala2
