+++ core/src/main/scala/chisel3/Printable.scala
@@ -65,7 +65,7 @@ object Printable {
 
   /** Pack standard printf fmt, args* style into Printable
     */
-  def pack(fmt: String, data: Data*): Printable = {
+  def pack(fmt: String, data: DataImpl*): Printable = {
     val args = data.iterator
     // Error handling
     def carrotAt(index: Int) = (" " * index) + "^"
