+++ core/src/main/scala/chisel3/Printf.scala
@@ -5,14 +5,16 @@ package chisel3
 import chisel3.internal._
 import chisel3.internal.Builder.pushCommand
 import chisel3.experimental.SourceInfo
-import scala.language.experimental.macros
-import scala.reflect.macros.blackbox
+import chisel3.PrintfMacrosCompat._
 
 /** Prints a message in simulation
   *
   * See apply methods for use
   */
 object printf {
+
+  /** Named class for [[printf]]s. */
+  final class Printf private[chisel3] (val pable: Printable) extends VerificationStatement
 
   /** Helper for packing escape characters */
   private[chisel3] def format(formatIn: String): String = {
@@ -35,41 +37,6 @@ object printf {
     }
     formatIn.map(escaped).mkString("")
   }
-
-  private[chisel3] def _checkFormatString(c: blackbox.Context)(fmt: c.Tree): Unit = {
-    import c.universe._
-
-    val errorString = "The s-interpolator prints the Scala .toString of Data objects rather than the value " +
-      "of the hardware wire during simulation. Use the cf-interpolator instead. If you want " +
-      "an elaboration time print, use println."
-
-    // Error on Data in the AST by matching on the Scala 2.13 string
-    // interpolation lowering to concatenation
-    def throwOnChiselData(x: c.Tree): Unit = x match {
-      case q"$x+$y" => {
-        if (x.tpe <:< typeOf[chisel3.Data] || y.tpe <:< typeOf[chisel3.Data]) {
-          c.error(c.enclosingPosition, errorString)
-        } else {
-          throwOnChiselData(x)
-          throwOnChiselData(y)
-        }
-      }
-      case _ =>
-    }
-    throwOnChiselData(fmt)
-
-    fmt match {
-      case q"scala.StringContext.apply(..$_).s(..$_)" =>
-        c.error(
-          c.enclosingPosition,
-          errorString
-        )
-      case _ =>
-    }
-  }
-
-  /** Named class for [[printf]]s. */
-  final class Printf private[chisel3] (val pable: Printable) extends VerificationStatement
 
   /** Prints a message in simulation
     *
@@ -109,19 +76,7 @@ object printf {
     * @param data format string varargs containing data to print
     */
   def apply(fmt: String, data: Bits*)(implicit sourceInfo: SourceInfo): Printf =
-    macro _applyMacroWithInterpolatorCheck
-
-  def _applyMacroWithInterpolatorCheck(
-    c:          blackbox.Context
-  )(fmt:        c.Tree,
-    data:       c.Tree*
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    _checkFormatString(c)(fmt)
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("printfWithReset"))
-    q"$apply_impl_do(_root_.chisel3.Printable.pack($fmt, ..$data))($sourceInfo)"
-  }
+    macro PrintfMacrosCompat._applyMacroWithInterpolatorCheck
 
   /** Prints a message in simulation
     *
