+++ core/src/main/scala/chisel3/VerificationStatement.scala
@@ -9,10 +9,11 @@ import chisel3.internal.firrtl.ir._
 import chisel3.layer.block
 import chisel3.layers
 import chisel3.util.circt.IfElseFatalIntrinsic
-import scala.annotation.nowarn
-import scala.language.experimental.macros
-import scala.reflect.macros.blackbox
-import scala.reflect.macros.blackbox.Context
+
+/** Base class for all verification statements: Assert, Assume, Cover, Stop and Printf. */
+abstract class VerificationStatement extends NamedComponent {
+  _parent.foreach(_.addId(this))
+}
 
 /** Scaladoc information for internal verification statement macros
   * that are used in objects assert, assume and cover.
@@ -29,6 +30,9 @@ import scala.reflect.macros.blackbox.Context
 trait VerifPrintMacrosDoc
 
 object assert extends VerifPrintMacrosDoc {
+
+  /** Named class for assertions. */
+  final class Assert private[chisel3] () extends VerificationStatement
 
   /** Checks for a condition to be valid in the circuit at rising clock edge
     * when not in reset. If the condition evaluates to false, the circuit
@@ -50,7 +54,7 @@ object assert extends VerifPrintMacrosDoc {
     data:    Bits*
   )(
     implicit sourceInfo: SourceInfo
-  ): Assert = macro _applyMacroWithInterpolatorCheck
+  ): Assert = macro VerifStmtMacrosCompat.assert._applyMacroWithInterpolatorCheck
 
   /** Checks for a condition to be valid in the circuit at all times. If the
     * condition evaluates to false, the circuit simulation stops with an error.
@@ -71,71 +75,18 @@ object assert extends VerifPrintMacrosDoc {
     message: Printable
   )(
     implicit sourceInfo: SourceInfo
-  ): Assert = macro _applyMacroWithPrintableMessage
+  ): Assert = macro VerifStmtMacrosCompat.assert._applyMacroWithPrintableMessage
 
   def apply(cond: Bool)(implicit sourceInfo: SourceInfo): Assert =
-    macro _applyMacroWithNoMessage
+    macro VerifStmtMacrosCompat.assert._applyMacroWithNoMessage
 
   import VerificationStatement._
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithInterpolatorCheck(
-    c:          blackbox.Context
-  )(cond:       c.Tree,
-    message:    c.Tree,
-    data:       c.Tree*
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    printf._checkFormatString(c)(message)
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
-    q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.Some(_root_.chisel3.Printable.pack($message, ..$data)))($sourceInfo)"
-  }
 
   /** An elaboration-time assertion. Calls the built-in Scala assert function. */
   def apply(cond: Boolean, message: => String): Unit = Predef.assert(cond, message)
 
   /** An elaboration-time assertion. Calls the built-in Scala assert function. */
   def apply(cond: Boolean): Unit = Predef.assert(cond, "")
-
-  /** Named class for assertions. */
-  final class Assert private[chisel3] () extends VerificationStatement
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithStringMessage(
-    c:          blackbox.Context
-  )(cond:       c.Tree,
-    message:    c.Tree,
-    data:       c.Tree*
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
-    q"$apply_impl_do($cond, ${getLine(c)},_root_.scala.Some(_root_.chisel3.Printable.pack($message,..$data)))($sourceInfo)"
-  }
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithPrintableMessage(
-    c:          blackbox.Context
-  )(cond:       c.Tree,
-    message:    c.Tree
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
-    q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.Some($message))($sourceInfo)"
-  }
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithNoMessage(
-    c:          blackbox.Context
-  )(cond:       c.Tree
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
-    q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.None)($sourceInfo)"
-  }
 
   private def emitIfElseFatalIntrinsic(
     clock:     Clock,
@@ -184,6 +135,9 @@ object assert extends VerifPrintMacrosDoc {
 
 object assume extends VerifPrintMacrosDoc {
 
+  /** Named class for assumptions. */
+  final class Assume private[chisel3] () extends VerificationStatement
+
   /** Assumes a condition to be valid in the circuit at all times.
     * Acts like an assertion in simulation and imposes a declarative
     * assumption on the state explored by formal tools.
@@ -209,7 +163,7 @@ object assume extends VerifPrintMacrosDoc {
     data:    Bits*
   )(
     implicit sourceInfo: SourceInfo
-  ): Assume = macro _applyMacroWithInterpolatorCheck
+  ): Assume = macro VerifStmtMacrosCompat.assume._applyMacroWithInterpolatorCheck
 
   /** Assumes a condition to be valid in the circuit at all times.
     * Acts like an assertion in simulation and imposes a declarative
@@ -229,10 +183,10 @@ object assume extends VerifPrintMacrosDoc {
     message: Printable
   )(
     implicit sourceInfo: SourceInfo
-  ): Assume = macro _applyMacroWithPrintableMessage
+  ): Assume = macro VerifStmtMacrosCompat.assume._applyMacroWithPrintableMessage
 
   def apply(cond: Bool)(implicit sourceInfo: SourceInfo): Assume =
-    macro _applyMacroWithNoMessage
+    macro VerifStmtMacrosCompat.assume._applyMacroWithNoMessage
 
   /** An elaboration-time assumption. Calls the built-in Scala assume function. */
   def apply(cond: Boolean, message: => String): Unit = Predef.assume(cond, message)
@@ -240,60 +194,7 @@ object assume extends VerifPrintMacrosDoc {
   /** An elaboration-time assumption. Calls the built-in Scala assume function. */
   def apply(cond: Boolean): Unit = Predef.assume(cond, "")
 
-  /** Named class for assumptions. */
-  final class Assume private[chisel3] () extends VerificationStatement
-
   import VerificationStatement._
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithInterpolatorCheck(
-    c:          blackbox.Context
-  )(cond:       c.Tree,
-    message:    c.Tree,
-    data:       c.Tree*
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    printf._checkFormatString(c)(message)
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
-    q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.Some(_root_.chisel3.Printable.pack($message, ..$data)))($sourceInfo)"
-  }
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithStringMessage(
-    c:          blackbox.Context
-  )(cond:       c.Tree,
-    message:    c.Tree,
-    data:       c.Tree*
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
-    q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.Some(_root_.chisel3.Printable.pack($message, ..$data)))($sourceInfo)"
-  }
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithPrintableMessage(
-    c:          blackbox.Context
-  )(cond:       c.Tree,
-    message:    c.Tree
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
-    q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.Some($message))($sourceInfo)"
-  }
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithNoMessage(
-    c:          blackbox.Context
-  )(cond:       c.Tree
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLinePrintable"))
-    q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.None)($sourceInfo)"
-  }
 
   /** This will be removed in Chisel 3.6 in favor of the Printable version
     *
@@ -339,6 +240,11 @@ object assume extends VerifPrintMacrosDoc {
 
 object cover extends VerifPrintMacrosDoc {
 
+  /** Named class for cover statements. */
+  final class Cover private[chisel3] () extends VerificationStatement
+
+  type SourceLineInfo = (String, Int)
+
   /** Declares a condition to be covered.
     * At ever clock event, a counter is incremented iff the condition is active
     * and reset is inactive.
@@ -352,37 +258,9 @@ object cover extends VerifPrintMacrosDoc {
     */
   // Macros currently can't take default arguments, so we need two functions to emulate defaults.
   def apply(cond: Bool, message: String)(implicit sourceInfo: SourceInfo): Cover =
-    macro _applyMacroWithMessage
+    macro VerifStmtMacrosCompat.cover._applyMacroWithMessage
   def apply(cond: Bool)(implicit sourceInfo: SourceInfo): Cover =
-    macro _applyMacroWithNoMessage
-
-  /** Named class for cover statements. */
-  final class Cover private[chisel3] () extends VerificationStatement
-
-  import VerificationStatement._
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithNoMessage(
-    c:          blackbox.Context
-  )(cond:       c.Tree
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLine"))
-    q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.None)($sourceInfo)"
-  }
-
-  /** @group VerifPrintMacros */
-  def _applyMacroWithMessage(
-    c:          blackbox.Context
-  )(cond:       c.Tree,
-    message:    c.Tree
-  )(sourceInfo: c.Tree
-  ): c.Tree = {
-    import c.universe._
-    val apply_impl_do = symbolOf[this.type].asClass.module.info.member(TermName("_applyWithSourceLine"))
-    q"$apply_impl_do($cond, ${getLine(c)}, _root_.scala.Some($message))($sourceInfo)"
-  }
+    macro VerifStmtMacrosCompat.cover._applyMacroWithNoMessage
 
   /** @group VerifPrintMacros */
   def _applyWithSourceLine(
@@ -436,20 +314,10 @@ object stop {
   }
 }
 
-/** Base class for all verification statements: Assert, Assume, Cover, Stop and Printf. */
-abstract class VerificationStatement extends NamedComponent {
-  _parent.foreach(_.addId(this))
-}
-
 /** Helper functions for common functionality required by stop, assert, assume or cover */
 private object VerificationStatement {
 
   type SourceLineInfo = (String, Int)
-
-  def getLine(c: blackbox.Context): SourceLineInfo = {
-    val p = c.enclosingPosition
-    (p.source.file.name, p.line): @nowarn // suppress, there's no clear replacement
-  }
 
   def formatFailureMessage(
     kind:     String,
